Rename stores and delegate node creation
