+++ src/components/LayersToolbar.vue
@@ -6,10 +6,10 @@
         <button @click="onAddGroup" title="Add group" class="p-1 rounded-md border border-white/15 bg-white/5 hover:bg-white/10">
           <img :src="toolbarIcons.group" alt="Add group" class="w-4 h-4">
         </button>
-        <button @click="onCopy" :disabled="!layers.selectionExists" title="Copy layer" class="p-1 rounded-md border border-white/15 bg-white/5 hover:bg-white/10 disabled:opacity-50 disabled:cursor-not-allowed">
+        <button @click="onCopy" :disabled="!nodes.layerSelectionExists" title="Copy layer" class="p-1 rounded-md border border-white/15 bg-white/5 hover:bg-white/10 disabled:opacity-50 disabled:cursor-not-allowed">
           <img :src="toolbarIcons.copy" alt="Copy layer" class="w-4 h-4">
         </button>
-        <button @click="onMerge" :disabled="layers.selectionCount < 2" title="Merge layers" class="p-1 rounded-md border border-white/15 bg-white/5 hover:bg-white/10 disabled:opacity-50 disabled:cursor-not-allowed">
+        <button @click="onMerge" :disabled="nodes.selectedLayerCount < 2" title="Merge layers" class="p-1 rounded-md border border-white/15 bg-white/5 hover:bg-white/10 disabled:opacity-50 disabled:cursor-not-allowed">
           <img :src="toolbarIcons.merge" alt="Merge layers" class="w-4 h-4">
         </button>
         <button @click="onSplit" :disabled="!canSplit" title="Split disconnected" class="p-1 rounded-md border border-white/15 bg-white/5 hover:bg-white/10 disabled:opacity-50 disabled:cursor-not-allowed">
@@ -27,33 +27,33 @@ import { useService } from '../services';
 import { computed } from 'vue';
 import toolbarIcons from '../image/layer_toolbar';
 
-const { layers, output } = useStore();
+const { nodeTree, nodes, output } = useStore();
 const { layerTool: layerSvc, layerPanel, query } = useService();
 
-const hasEmptyLayers = computed(() => layers.order.some(id => layers.getProperty(id, 'pixels').length === 0));
-const canSplit = computed(() => layers.selectedIds.some(id => layers.disconnectedCountOf(id) > 1));
+const hasEmptyLayers = computed(() => nodeTree.layerOrder.some(id => (nodes.getProperty(id, 'pixels') || []).length === 0));
+const canSplit = computed(() => nodeTree.selectedLayerIds.some(id => nodes.disconnectedCountOfLayer(id) > 1));
 
 const onAdd = () => {
     output.setRollbackPoint();
-    const above = layers.selectionCount ? query.uppermost(layers.selectedIds) : null;
-    const id = layers.createLayer({});
-    layers.insert([id], above, false);
-    layers.replaceSelection([id]);
+    const above = nodeTree.selectedLayerCount ? query.uppermost(nodeTree.selectedLayerIds) : null;
+    const id = nodes.createLayer({});
+    nodeTree.insert([id], above, false);
+    nodeTree.replaceSelection([id]);
     layerPanel.setScrollRule({ type: 'follow', target: id });
     output.commit();
 };
 const onAddGroup = () => {
     output.setRollbackPoint();
-    const selected = layers.selectedNodeIds;
-    const id = layers.createGroup({});
+    const selected = nodeTree.selectedNodeIds;
+    const id = nodes.createGroup({});
     if (selected.length === 0) {
-        layers.putIn([id], null, false);
+        nodeTree.putIn([id], null, false);
     } else {
         const lowermost = selected[0];
-        layers.insert([id], lowermost, true);
-        layers.putIn(selected, id, true);
+        nodeTree.insert([id], lowermost, true);
+        nodeTree.putIn(selected, id, true);
     }
-    layers.replaceSelection([id]);
+    nodeTree.replaceSelection([id]);
     layerPanel.setRange(id, id);
     layerPanel.setScrollRule({ type: 'follow', target: id });
     output.commit();
@@ -61,20 +61,20 @@ const onAddGroup = () => {
 const onMerge = () => {
     output.setRollbackPoint();
     const id = layerSvc.mergeSelected();
-    layers.replaceSelection([id]);
+    nodeTree.replaceSelection([id]);
     layerPanel.setScrollRule({ type: 'follow', target: id });
     output.commit();
 };
 const onCopy = () => {
     output.setRollbackPoint();
     const ids = layerSvc.copySelected();
-    layers.replaceSelection(ids);
+    nodeTree.replaceSelection(ids);
     layerPanel.setScrollRule({ type: 'follow', target: ids[0] });
     output.commit();
 };
 const onSelectEmpty = () => {
     const ids = query.empty();
-    layers.replaceSelection(ids);
+    nodeTree.replaceSelection(ids);
     layerPanel.setScrollRule({ type: 'follow', target: ids[0] });
 };
 const onSplit = () => {
