+++ src/components/ViewportInfo.vue
@@ -14,11 +14,11 @@ import { useStore } from '../stores';
 import { useService } from '../services';
 import { getPixelUnion, rgbaCssU32, rgbaCssObj } from '../utils';
 
-const { viewport: viewportStore, layers, input } = useStore();
+const { viewport: viewportStore, nodeTree, nodes, input } = useStore();
 const { toolSelection: toolSelectionService } = useService();
 
 const selectedAreaPixelCount = computed(() => {
-    const pixels = getPixelUnion(layers.getProperties(layers.selectedIds));
+    const pixels = getPixelUnion(nodes.getProperties(nodeTree.selectedLayerIds));
     return pixels.length;
   });
 
@@ -30,7 +30,7 @@ const pixelInfo = computed(() => {
       const colorObject = input.readPixel(coord);
       return `[${px},${py}] ${rgbaCssObj(colorObject)}`;
     } else {
-      const colorU32 = layers.compositeColorAt(coord);
+      const colorU32 = nodes.compositeColorAt(coord);
       return `[${px},${py}] ${rgbaCssU32(colorU32)}`;
     }
   });
