+++ src/services/layerPanel.js
@@ -3,7 +3,7 @@ import { reactive, toRefs, watch, computed } from 'vue';
 import { useStore } from '../stores';
 
 export const useLayerPanelService = defineStore('layerPanelService', () => {
-    const { layers } = useStore();
+    const { nodeTree } = useStore();
 
     const state = reactive({
         anchorId: null,
@@ -19,10 +19,8 @@ export const useLayerPanelService = defineStore('layerPanelService', () => {
     function enableWatch() {
         if (!stopWatch) {
             stopWatch = watch(
-                () => layers.selectedIds,
-                () => {
-                    clearRange();
-                },
+                () => nodeTree.selectedLayerIds,
+                () => { clearRange(); },
                 { flush: 'sync' }
             );
         }
@@ -37,9 +35,9 @@ export const useLayerPanelService = defineStore('layerPanelService', () => {
     function dfs(skipFolded = false) {
         const order = [];
         const ancestors = new Map();
-        const walk = (nodes, anc) => {
-            for (let i = nodes.length - 1; i >= 0; i--) {
-                const node = nodes[i];
+        const walk = (list, anc) => {
+            for (let i = list.length - 1; i >= 0; i--) {
+                const node = list[i];
                 ancestors.set(node.id, anc.slice());
                 order.push(node.id);
                 if (node.children && !(skipFolded && folded[node.id])) {
@@ -47,15 +45,15 @@ export const useLayerPanelService = defineStore('layerPanelService', () => {
                 }
             }
         };
-        walk(layers.tree, []);
+        walk(nodeTree.tree, []);
         return { order, ancestors };
     }
 
     function visibleAncestor(id, orderSet) {
-        let info = layers._findNode(id);
+        let info = nodeTree._findNode(id);
         while (info && !orderSet.has(info.node.id)) {
             if (!info.parent) return null;
-            info = layers._findNode(info.parent.id);
+            info = nodeTree._findNode(info.parent.id);
         }
         return info ? info.node.id : null;
     }
@@ -65,7 +63,7 @@ export const useLayerPanelService = defineStore('layerPanelService', () => {
         if (anchorId == null || tailId == null) {
             state.anchorId = null;
             state.tailId = null;
-            layers.clearSelection();
+            nodeTree.clearSelection();
             return;
         }
 
@@ -75,7 +73,7 @@ export const useLayerPanelService = defineStore('layerPanelService', () => {
         if (idxA === -1 || idxB === -1) {
             state.anchorId = null;
             state.tailId = null;
-            layers.clearSelection();
+            nodeTree.clearSelection();
             return;
         }
         const [start, end] = idxA < idxB ? [idxA, idxB] : [idxB, idxA];
@@ -85,7 +83,7 @@ export const useLayerPanelService = defineStore('layerPanelService', () => {
             ...(ancestors.get(tailId) || []),
         ]);
         const selection = slice.filter(id => !ancToRemove.has(id));
-        layers.replaceSelection(selection);
+        nodeTree.replaceSelection(selection);
         state.anchorId = anchorId;
         state.tailId = tailId;
         enableWatch();
@@ -105,7 +103,7 @@ export const useLayerPanelService = defineStore('layerPanelService', () => {
         if (event.shiftKey) {
             setRange(state.anchorId ?? id, id);
         } else if (event.ctrlKey || event.metaKey) {
-            layers.toggleSelection(id);
+            nodeTree.toggleSelection(id);
         } else {
             setRange(id, id);
         }
@@ -113,12 +111,12 @@ export const useLayerPanelService = defineStore('layerPanelService', () => {
     }
 
     function onArrowUp(shift, ctrl) {
-        if (!layers.exists || ctrl) return;
+        if (!nodeTree.exists || ctrl) return;
         const { order } = dfs(true);
         if (!order.length) return;
         const orderSet = new Set(order);
         if (shift) {
-            if (!layers.selectionExists) return;
+            if (!nodeTree.layerSelectionExists) return;
             const tailVis = visibleAncestor(state.tailId, orderSet);
             const anchorVis = visibleAncestor(state.anchorId, orderSet);
             if (tailVis == null || anchorVis == null) return;
@@ -137,12 +135,12 @@ export const useLayerPanelService = defineStore('layerPanelService', () => {
     }
 
     function onArrowDown(shift, ctrl) {
-        if (!layers.exists || ctrl) return;
+        if (!nodeTree.exists || ctrl) return;
         const { order } = dfs(true);
         if (!order.length) return;
         const orderSet = new Set(order);
         if (shift) {
-            if (!layers.selectionExists) return;
+            if (!nodeTree.layerSelectionExists) return;
             const tailVis = visibleAncestor(state.tailId, orderSet);
             const anchorVis = visibleAncestor(state.anchorId, orderSet);
             if (tailVis == null || anchorVis == null) return;
