+++ src/services/tools.js
@@ -12,7 +12,7 @@ export const useDrawToolService = defineStore('drawToolService', () => {
     const overlayService = useOverlayService();
     const overlay = overlayService.addOverlay('draw');
     overlay.config.value = OVERLAY_CONFIG.ADD;
-    const { layers } = useStore();
+    const { nodeTree, nodes } = useStore();
     watch(() => tool.prepared === 'draw', (isDraw) => {
         if (!isDraw) {
             overlay.clear();
@@ -25,9 +25,9 @@ export const useDrawToolService = defineStore('drawToolService', () => {
         overlay.setPixels(pixel ? [pixel] : []);
     });
     watch(() => tool.dragPixel, (pixel) => {
-        if (tool.prepared !== 'draw' || layers.selectionCount !== 1) return;
-        const sourceId = layers.selectedIds[0];
-        if (layers.getProperty(sourceId, 'locked')) {
+        if (tool.prepared !== 'draw' || nodeTree.selectedLayerCount !== 1) return;
+        const sourceId = nodeTree.selectedLayerIds[0];
+        if (nodes.getProperty(sourceId, 'locked')) {
             if (pixel)
                 tool.setCursor({ stroke: CURSOR_CONFIG.LOCKED, rect: CURSOR_CONFIG.LOCKED });
             else
@@ -36,14 +36,14 @@ export const useDrawToolService = defineStore('drawToolService', () => {
         }
     });
     watch(() => tool.previewPixels, (pixels) => {
-        if (tool.prepared !== 'draw' || layers.selectionCount !== 1) return;
+        if (tool.prepared !== 'draw' || nodeTree.selectedLayerCount !== 1) return;
         overlay.setPixels(pixels);
     });
     watch(() => tool.affectedPixels, (pixels) => {
-        if (tool.prepared !== 'draw' || layers.selectionCount !== 1) return;
-        const id = layers.selectedIds[0];
-        if (layers.getProperty(id, 'locked')) return;
-        layers.addPixels(id, pixels);
+        if (tool.prepared !== 'draw' || nodeTree.selectedLayerCount !== 1) return;
+        const id = nodeTree.selectedLayerIds[0];
+        if (nodes.getProperty(id, 'locked')) return;
+        nodes.addPixelsToLayer(id, pixels);
     });
     return {};
 });
@@ -53,7 +53,7 @@ export const useEraseToolService = defineStore('eraseToolService', () => {
     const overlayService = useOverlayService();
     const overlay = overlayService.addOverlay('erase');
     overlay.config.value = OVERLAY_CONFIG.REMOVE;
-    const { layers } = useStore();
+    const { nodeTree, nodes } = useStore();
     watch(() => tool.prepared === 'erase', (isErase) => {
         if (!isErase) {
             overlay.clear();
@@ -66,10 +66,10 @@ export const useEraseToolService = defineStore('eraseToolService', () => {
         overlay.setPixels(pixel ? [pixel] : []);
     });
     watch(() => tool.dragPixel, (pixel) => {
-        if (tool.prepared !== 'erase' || layers.selectionCount !== 1) return;
-        const sourceId = layers.selectedIds[0];
-        if (layers.getProperty(sourceId, 'locked')) {
-            const sourceKeys = new Set(layers.getProperty(sourceId, 'pixels').map(coordToKey));
+        if (tool.prepared !== 'erase' || nodeTree.selectedLayerCount !== 1) return;
+        const sourceId = nodeTree.selectedLayerIds[0];
+        if (nodes.getProperty(sourceId, 'locked')) {
+            const sourceKeys = new Set((nodes.getProperty(sourceId, 'pixels') || []).map(coordToKey));
             if (pixel && sourceKeys.has(coordToKey(pixel)))
                 tool.setCursor({ stroke: CURSOR_CONFIG.LOCKED, rect: CURSOR_CONFIG.LOCKED });
             else
@@ -77,14 +77,14 @@ export const useEraseToolService = defineStore('eraseToolService', () => {
         }
     });
     watch(() => tool.previewPixels, (pixels) => {
-        if (tool.prepared !== 'erase' || layers.selectionCount !== 1) return;
+        if (tool.prepared !== 'erase' || nodeTree.selectedLayerCount !== 1) return;
         overlay.setPixels(pixels);
     });
     watch(() => tool.affectedPixels, (pixels) => {
-        if (tool.prepared !== 'erase' || layers.selectionCount !== 1) return;
-        const id = layers.selectedIds[0];
-        if (layers.getProperty(id, 'locked')) return;
-        layers.removePixels(id, pixels);
+        if (tool.prepared !== 'erase' || nodeTree.selectedLayerCount !== 1) return;
+        const id = nodeTree.selectedLayerIds[0];
+        if (nodes.getProperty(id, 'locked')) return;
+        nodes.removePixelsFromLayer(id, pixels);
     });
     return {};
 });
@@ -95,7 +95,7 @@ export const useCutToolService = defineStore('cutToolService', () => {
     const overlay = overlayService.addOverlay('cut');
     overlay.config.value = OVERLAY_CONFIG.REMOVE;
     const layerPanel = useLayerPanelService();
-    const { layers } = useStore();
+    const { nodeTree, nodes } = useStore();
     watch(() => tool.prepared === 'cut', (isCut) => {
         if (!isCut) {
             overlay.clear();
@@ -108,10 +108,10 @@ export const useCutToolService = defineStore('cutToolService', () => {
         overlay.setPixels(pixel ? [pixel] : []);
     });
     watch(() => tool.dragPixel, (pixel) => {
-        if (tool.prepared !== 'cut' || layers.selectionCount !== 1) return;
-        const sourceId = layers.selectedIds[0];
-        if (layers.getProperty(sourceId, 'locked')) {
-            const sourceKeys = new Set(layers.getProperty(sourceId, 'pixels').map(coordToKey));
+        if (tool.prepared !== 'cut' || nodeTree.selectedLayerCount !== 1) return;
+        const sourceId = nodeTree.selectedLayerIds[0];
+        if (nodes.getProperty(sourceId, 'locked')) {
+            const sourceKeys = new Set((nodes.getProperty(sourceId, 'pixels') || []).map(coordToKey));
             if (pixel && sourceKeys.has(coordToKey(pixel)))
                 tool.setCursor({ stroke: CURSOR_CONFIG.LOCKED, rect: CURSOR_CONFIG.LOCKED });
             else
@@ -119,14 +119,14 @@ export const useCutToolService = defineStore('cutToolService', () => {
         }
     });
     watch(() => tool.previewPixels, (pixels) => {
-        if (tool.prepared !== 'cut' || layers.selectionCount !== 1) return;
+        if (tool.prepared !== 'cut' || nodeTree.selectedLayerCount !== 1) return;
         overlay.setPixels(pixels);
     });
     watch(() => tool.affectedPixels, (pixels) => {
-        if (tool.prepared !== 'cut' || layers.selectionCount !== 1) return;
-        const sourceId = layers.selectedIds[0];
-        if (layers.getProperty(sourceId, 'locked')) return;
-        const sourceKeys = new Set(layers.getProperty(sourceId, 'pixels').map(coordToKey));
+        if (tool.prepared !== 'cut' || nodeTree.selectedLayerCount !== 1) return;
+        const sourceId = nodeTree.selectedLayerIds[0];
+        if (nodes.getProperty(sourceId, 'locked')) return;
+        const sourceKeys = new Set((nodes.getProperty(sourceId, 'pixels') || []).map(coordToKey));
 
         const cutCoords = [];
         const cutKeys = new Set();
@@ -140,17 +140,17 @@ export const useCutToolService = defineStore('cutToolService', () => {
 
         if (!cutCoords.length || cutKeys.size === sourceKeys.size) return;
 
-        layers.removePixels(sourceId, cutCoords);
-        const id = layers.createLayer({
-            name: `Cut of ${layers.getProperty(sourceId, 'name')}`,
-            color: layers.getProperty(sourceId, 'color'),
-            visibility: layers.getProperty(sourceId, 'visibility'),
+        nodes.removePixelsFromLayer(sourceId, cutCoords);
+        const id = nodes.createLayer({
+            name: `Cut of ${nodes.getProperty(sourceId, 'name')}`,
+            color: nodes.getProperty(sourceId, 'color'),
+            visibility: nodes.getProperty(sourceId, 'visibility'),
             pixels: cutCoords,
-            attributes: layers.getProperty(sourceId, 'attributes'),
+            attributes: nodes.getProperty(sourceId, 'attributes'),
         });
-        layers.insert([id], sourceId, false);
+        nodeTree.insert([id], sourceId, false);
 
-        layers.replaceSelection([sourceId]);
+        nodeTree.replaceSelection([sourceId]);
         layerPanel.setScrollRule({ type: 'follow', target: sourceId });
     });
     return {};
@@ -162,7 +162,7 @@ export const useTopToolService = defineStore('topToolService', () => {
     const overlay = overlayService.addOverlay('top');
     overlay.config.value = OVERLAY_CONFIG.ADD;
     const layerPanel = useLayerPanelService();
-    const { layers } = useStore();
+    const { nodeTree, nodes } = useStore();
     watch(() => tool.prepared === 'top', (isTop) => {
         if (!isTop) {
             overlay.clear();
@@ -176,8 +176,8 @@ export const useTopToolService = defineStore('topToolService', () => {
             overlay.clear();
             return;
         }
-        const id = layers.topVisibleIdAt(pixel);
-        if (id && layers.getProperty(id, 'locked')) {
+        const id = nodes.topVisibleIdAt(pixel);
+        if (id && nodes.getProperty(id, 'locked')) {
             tool.setCursor({ stroke: CURSOR_CONFIG.LOCKED, rect: CURSOR_CONFIG.LOCKED });
         }
         else {
@@ -187,14 +187,14 @@ export const useTopToolService = defineStore('topToolService', () => {
     });
     watch(() => tool.dragPixel, (pixel) => {
         if (tool.prepared !== 'top' || !pixel) return;
-        const id = layers.topVisibleIdAt(pixel);
+        const id = nodes.topVisibleIdAt(pixel);
         if (!id) return;
-        if (layers.getProperty(id, 'locked')) {
+        if (nodes.getProperty(id, 'locked')) {
             tool.setCursor({ stroke: CURSOR_CONFIG.LOCKED, rect: CURSOR_CONFIG.LOCKED });
         }
         else {
-            layers.insert([id], layers.idsTopToBottom[0], false);
-            layers.replaceSelection([id]);
+            nodeTree.insert([id], nodeTree.layerIdsTopToBottom[0], false);
+            nodeTree.replaceSelection([id]);
             layerPanel.setScrollRule({ type: 'follow', target: id });
             tool.setCursor({ stroke: CURSOR_CONFIG.TOP, rect: CURSOR_CONFIG.TOP });
         } 
@@ -208,7 +208,7 @@ export const useSelectService = defineStore('selectService', () => {
     const overlay = overlayService.addOverlay('select');
     overlay.config.value = OVERLAY_CONFIG.ADD;
     const layerPanel = useLayerPanelService();
-    const { layers, viewportEvent: viewportEvents } = useStore();
+    const { nodeTree, nodes, viewportEvent: viewportEvents } = useStore();
     let mode = 'select';
     watch(() => tool.prepared === 'select', (isSelect) => {
         if (!isSelect) {
@@ -223,12 +223,12 @@ export const useSelectService = defineStore('selectService', () => {
             overlay.clear();
             return;
         }
-        const id = layers.topVisibleIdAt(pixel);
+        const id = nodes.topVisibleIdAt(pixel);
         if (!viewportEvents.isPressed('Shift')) {
             mode = 'select';
             overlay.config.value = OVERLAY_CONFIG.ADD;
             tool.setCursor({ stroke: CURSOR_CONFIG.ADD_STROKE, rect: CURSOR_CONFIG.ADD_RECT });
-        } else if (layers.isSelected(id)) {
+        } else if (nodeTree.isSelected(id)) {
             mode = 'remove';
             overlay.config.value = OVERLAY_CONFIG.REMOVE;
             tool.setCursor({ stroke: CURSOR_CONFIG.REMOVE_STROKE, rect: CURSOR_CONFIG.REMOVE_RECT });
@@ -238,7 +238,7 @@ export const useSelectService = defineStore('selectService', () => {
             tool.setCursor({ stroke: CURSOR_CONFIG.ADD_STROKE, rect: CURSOR_CONFIG.ADD_RECT });
         }
 
-        if (id && layers.getProperty(id, 'locked')) {
+        if (id && nodes.getProperty(id, 'locked')) {
             tool.setCursor({ stroke: CURSOR_CONFIG.LOCKED, rect: CURSOR_CONFIG.LOCKED });
         }
         overlay.setLayers(id ? [id] : []);
@@ -246,8 +246,8 @@ export const useSelectService = defineStore('selectService', () => {
     watch(() => tool.dragPixel, (pixel) => {
         if (tool.prepared !== 'select') return;
         if (pixel) {
-            const id = layers.topVisibleIdAt(pixel);
-            if (id && layers.getProperty(id, 'locked')) {
+            const id = nodes.topVisibleIdAt(pixel);
+            if (id && nodes.getProperty(id, 'locked')) {
                 tool.setCursor({ stroke: CURSOR_CONFIG.LOCKED, rect: CURSOR_CONFIG.LOCKED });
                 return;
             }
@@ -258,14 +258,14 @@ export const useSelectService = defineStore('selectService', () => {
         if (tool.prepared !== 'select') return;
         const intersectedIds = [];
         for (const coord of pixels) {
-            const id = layers.topVisibleIdAt(coord);
+            const id = nodes.topVisibleIdAt(coord);
             if (id === null) continue;
-            if (!layers.getProperty(id, 'locked')) intersectedIds.push(id);
+            if (!nodes.getProperty(id, 'locked')) intersectedIds.push(id);
         }
         const highlightIds = [];
         intersectedIds.forEach(id => {
-            if (mode === 'remove' && !layers.isSelected(id)) return;
-            if (mode === 'add' && layers.isSelected(id)) return;
+            if (mode === 'remove' && !nodeTree.isSelected(id)) return;
+            if (mode === 'add' && nodeTree.isSelected(id)) return;
             highlightIds.push(id);
         });
         overlay.setLayers(highlightIds);
@@ -275,10 +275,10 @@ export const useSelectService = defineStore('selectService', () => {
         if (pixels.length > 0) {
             const intersectedIds = new Set();
             for (const coord of pixels) {
-                const id = layers.topVisibleIdAt(coord);
-                if (id !== null && !layers.getProperty(id, 'locked')) intersectedIds.add(id);
+            const id = nodes.topVisibleIdAt(coord);
+            if (id !== null && !nodes.getProperty(id, 'locked')) intersectedIds.add(id);
             }
-            const currentSelection = new Set(mode === 'select' ? [] : layers.selectedIds);
+            const currentSelection = new Set(mode === 'select' ? [] : nodeTree.selectedLayerIds);
             if (mode === 'add') {
                 intersectedIds.forEach(id => currentSelection.add(id));
             } else if (mode === 'remove') {
@@ -286,10 +286,10 @@ export const useSelectService = defineStore('selectService', () => {
             } else {
                 intersectedIds.forEach(id => currentSelection.add(id));
             }
-            layers.replaceSelection([...currentSelection]);
-            layerPanel.setScrollRule({ type: 'follow', target: layers.selectedIds[0] });
+            nodeTree.replaceSelection([...currentSelection]);
+            layerPanel.setScrollRule({ type: 'follow', target: nodeTree.selectedLayerIds[0] });
         } else if (mode === 'select') {
-            layers.clearSelection();
+            nodeTree.clearSelection();
         }
     });
     return {};
@@ -300,7 +300,7 @@ export const useGlobalEraseToolService = defineStore('globalEraseToolService', (
     const overlayService = useOverlayService();
     const overlay = overlayService.addOverlay('globalErase');
     overlay.config.value = OVERLAY_CONFIG.REMOVE;
-    const { layers } = useStore();
+    const { nodeTree, nodes } = useStore();
     watch(() => tool.prepared === 'globalErase', (isGlobalErase) => {
         if (!isGlobalErase) {
             overlay.clear();
@@ -315,9 +315,9 @@ export const useGlobalEraseToolService = defineStore('globalEraseToolService', (
     watch(() => tool.dragPixel, (pixel) => {
         if (tool.prepared !== 'globalErase') return;
         if (pixel){
-            const lockedIds = layers.order.filter(id => layers.getProperty(id, 'locked'));
+            const lockedIds = nodeTree.layerOrder.filter(id => nodes.getProperty(id, 'locked'));
             for (const id of lockedIds) {
-                const lockedPixels = new Set(layers.getProperty(id, 'pixels').map(coordToKey));
+                const lockedPixels = new Set((nodes.getProperty(id, 'pixels') || []).map(coordToKey));
                 if (lockedPixels.has(coordToKey(pixel))) {
                     tool.setCursor({ stroke: CURSOR_CONFIG.LOCKED, rect: CURSOR_CONFIG.LOCKED });
                     return;
@@ -330,10 +330,10 @@ export const useGlobalEraseToolService = defineStore('globalEraseToolService', (
         if (tool.prepared !== 'globalErase') return;
         const erasablePixels = [];
         if (pixels.length) {
-            const unlockedIds = layers.order.filter(id => !layers.getProperty(id, 'locked'));
+            const unlockedIds = nodeTree.layerOrder.filter(id => !nodes.getProperty(id, 'locked'));
             const unlockedPixels = new Set();
             for (const id of unlockedIds) {
-                layers.getProperty(id, 'pixels').forEach(coord => unlockedPixels.add(coordToKey(coord)));
+                (nodes.getProperty(id, 'pixels') || []).forEach(coord => unlockedPixels.add(coordToKey(coord)));
             }
             for (const coord of pixels) {
                 if (unlockedPixels.has(coordToKey(coord))) erasablePixels.push(coord);
@@ -343,15 +343,15 @@ export const useGlobalEraseToolService = defineStore('globalEraseToolService', (
     });
     watch(() => tool.affectedPixels, (pixels) => {
         if (tool.prepared !== 'globalErase' || !pixels.length) return;
-        const targetIds = (layers.selectionExists ? layers.selectedIds : layers.order)
-            .filter(id => !layers.getProperty(id, 'locked'));
+        const targetIds = (nodeTree.layerSelectionExists ? nodeTree.selectedLayerIds : nodeTree.layerOrder)
+            .filter(id => !nodes.getProperty(id, 'locked'));
         for (const id of targetIds) {
-            const targetKeys = new Set(layers.getProperty(id, "pixels").map(coordToKey));
+            const targetKeys = new Set((nodes.getProperty(id, "pixels") || []).map(coordToKey));
             const pixelsToRemove = [];
             for (const coord of pixels) {
                 if (targetKeys.has(coordToKey(coord))) pixelsToRemove.push(coord);
             }
-            if (pixelsToRemove.length) layers.removePixels(id, pixelsToRemove);
+            if (pixelsToRemove.length) nodes.removePixelsFromLayer(id, pixelsToRemove);
         }
     });
     return {};
